MRP-2024 add zmq augmentation loop to the refactory.
