+++ rtbkit/plugins/augmentor_interface/augmentor_interface.mk
@@ -1,5 +1,6 @@
 $(eval $(call library,http_augmentor,http_augmentor_interface.cc,rtb_router openrtb_bid_request))
+$(eval $(call library,zmq_augmentor,zmq_augmentor_interface.cc,rtb_router openrtb_bid_request))
 
-augmentor_interface_plugins: $(LIB)/libhttp_augmentor.so
+augmentor_interface_plugins: $(LIB)/libhttp_augmentor.so $(LIB)/libzmq_augmentor.so
 
 .PHONY: augmentor_interface_plugins
