+++ rtbkit/plugins/augmentor_interface/http_augmentor_interface.h
@@ -82,17 +82,6 @@ struct HttpAugmentorInterface : public AugmentorInterface
 
 private:
 
-    typedef boost::function<void (const std::shared_ptr<AugmentationInfo> &)>
-        OnFinished;
-
-    struct Entry {
-        std::shared_ptr<AugmentationInfo> info;
-        std::set<std::string> outstanding;
-        std::map<std::string, std::set<std::string> > augmentorAgents;
-        OnFinished onFinished;
-        Date timeout;
-    };
-
     /** List of auctions we're currently augmenting.  Once the augmentation
         process is finished the auction will be passed on.
     */
